SqliteGcbmResultsProvider now filters out year = 0 (spinup) if present in results.
Updated template _toc.yml for compatibility with latest jupyter-book - now incompatible with older jupyter-book 0.9.1.

Changed 'quiet' parameter in report plots to 'display' to more accurately reflect what it does: display=True, the default, causes the graph to be displayed immediately in the rendered report; display=False returns the graph object so that it can be generated in one place and displayed elsewhere (see tabbed_gcbm template).